+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * UI:
     * Made the "Tool Options" view more compact
     * Template system for new files
+    * Startup dialog
 * I/O:
     * More color formats are supported for the SVG importer
 * Scripting: